ticketCar/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.5
Django==5.0.4
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gcloud==0.18.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
mysqlclient==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-jwt==4.1.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
